requirements.txt
Django==4.2.2
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
drf-yasg==1.21.5
django-filter==23.2
langchain==0.0.224
faiss-cpu==1.7.4
openai==0.27.7
tiktoken==0.3.3
unstructured==0.7.1
tabulate==0.9.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
tokenizers==0.13.3
sec-api
fredapi
spacy
